Corrigido injeção de dependência em loop
